+++ sdk/monitor/Azure.Monitor.OpenTelemetry.Exporter/tests/Azure.Monitor.OpenTelemetry.Exporter.Demo/Logs/LogDemo.cs
@@ -17,7 +17,11 @@ namespace Azure.Monitor.OpenTelemetry.Exporter.Demo.Logs
             {
                 builder.AddOpenTelemetry(options =>
                 {
-                    options.AddAzureMonitorLogExporter(o => o.ConnectionString = connectionString, credential);
+                    options.AddAzureMonitorLogExporter(o =>
+                    {
+                        o.ConnectionString = connectionString;
+                        o.Credential = credential;
+                    });
                 });
             });
         }